everything works great now! fixed lack of accuracy changes problem.
- version 7 can be called done!
- fixed non-increasing accuracy problem
- interactive 3D scatterplot
- the scatterplots are now resized to the min and max of the obj values in each of the 3 dimensions
- every solution and its obj values for each run's each generation is saved to a txt file.
- random selection also implemented to later be compared to the NSGA2 algorithm.

there was a big problem. the bitstring generated by the minimize function had strings instead of ints, "1" instead of 1.

this would mess up the if conditions, cuz the ifs were like:
if bitstring[int(device.id)] == 1:
and it would never be true, cuz the output of bitstring[int(device.id)] is always string and never int, so i fixed it. both in the conditions and also replacing every "1" and "0" with 1 and 0 in the bitstring.